refactor: move *.js -> src/ (#107)
Related: https://github.com/liferay/liferay-npm-tools/issues/107
