+++ packages/liferay-jest-junit-reporter/package.json
@@ -2,7 +2,7 @@
 	"name": "liferay-jest-junit-reporter",
 	"version": "1.0.1",
 	"description": "A JUnit reporter for Jest and Liferay CI",
-	"main": "index.js",
+	"main": "src/index.js",
 	"scripts": {
 		"preversion": "cd ../.. && yarn ci",
 		"test": "jest"
