+++ packages/liferay-jest-junit-reporter/src/index.js
